+++ winery-spring-mvc/src/main/webapp/WEB-INF/jsp/wine/buy.jsp
@@ -30,7 +30,7 @@
     </table>
 
     <form:form method="POST"
-               action="${pageContext.request.contextPath}/wine/buyAmount/${wine.id}" modelAttribute="wineBuy">
+               action="${pageContext.request.contextPath}/user/wine/buyAmount/${wine.id}" modelAttribute="wineBuy">
          <div class="form-group">
              <form:label path="amount" cssClass="col-sm-2 control-label">Enter desired amount:</form:label>
              <div class="col-sm-2">
@@ -40,7 +40,7 @@
              <button class="btn btn-success" type="submit">Buy</button>
          </div>
 
-        <my:a href="/" class="btn btn-danger">Back</my:a>&emsp;
+        <my:a href="/user" class="btn btn-danger">Back</my:a>&emsp;
     </form:form>
 </jsp:attribute>
 </my:pagetemplate>
